refactor(desktop): split agent chat message card presenters
Move message-type renderers and card state derivation into focused modules so the main card stays a small composition layer that is easier to evolve and test.
