+++ apps/desktop/src/components/features/agents/agent-chat/agent-chat-message-card-content.tsx
@@ -1,0 +1,229 @@
+import { MarkdownRenderer } from "@/components/ui/markdown-renderer";
+import { cn } from "@/lib/utils";
+import type { AgentChatMessage } from "@/types/agent-orchestrator";
+import {
+  type AgentModelSelection,
+  type AgentRole,
+  isOdtWorkflowMutationToolName,
+} from "@openducktor/core";
+import { Brain, Hammer, MessageSquareQuote } from "lucide-react";
+import type { ReactElement } from "react";
+import {
+  SYSTEM_PROMPT_PREFIX,
+  getAssistantFooterData,
+  roleLabel,
+  toSingleLineMarkdown,
+} from "./agent-chat-message-card-model";
+import {
+  RegularToolMessage,
+  WorkflowToolMessage,
+  assistantRoleIcon,
+} from "./agent-chat-message-card-tool-presenters";
+
+export type MessageHeaderProps = {
+  message: AgentChatMessage;
+  sessionRole: AgentRole | null;
+  timeLabel: string;
+  showHeader: boolean;
+  assistantRole: AgentRole | null;
+  compactPadding: boolean;
+};
+
+export const MessageHeader = ({
+  message,
+  sessionRole,
+  timeLabel,
+  showHeader,
+  assistantRole,
+  compactPadding,
+}: MessageHeaderProps): ReactElement | null => {
+  if (!showHeader) {
+    return null;
+  }
+
+  return (
+    <header
+      className={cn(
+        "mb-1 flex items-center justify-between gap-2 text-[11px] font-semibold uppercase tracking-wide text-slate-500",
+        message.role === "assistant" ? "mb-2" : "mb-1",
+        compactPadding ? "" : "px-1",
+      )}
+    >
+      <span className="inline-flex items-center gap-1">
+        {message.role === "thinking" ? <Brain className="size-3" /> : null}
+        {message.role === "tool" ? <Hammer className="size-3" /> : null}
+        {message.role === "assistant" && assistantRole ? assistantRoleIcon(assistantRole) : null}
+        {roleLabel(message.role, sessionRole, message)}
+      </span>
+      {timeLabel ? <span className="font-normal normal-case">{timeLabel}</span> : null}
+    </header>
+  );
+};
+
+type ReasoningMessageProps = {
+  content: string;
+  completed: boolean;
+  timeLabel: string;
+};
+
+const ReasoningMessage = ({
+  content,
+  completed,
+  timeLabel,
+}: ReasoningMessageProps): ReactElement => {
+  if (completed) {
+    return (
+      <details className="px-1 py-0.5">
+        <summary className="flex min-h-6 cursor-pointer items-center gap-2 text-xs text-slate-700">
+          <Brain className="size-3.5 shrink-0 text-slate-500" />
+          <span className="shrink-0 font-medium text-slate-500">Thinking</span>
+          <span className="min-w-0 flex-1 truncate text-slate-600">
+            {toSingleLineMarkdown(content || "Reasoning complete")}
+          </span>
+          {timeLabel ? (
+            <span className="shrink-0 text-[11px] text-slate-500">{timeLabel}</span>
+          ) : null}
+        </summary>
+        <div className="pl-6 pt-2">
+          <MarkdownRenderer markdown={content || "Reasoning complete"} variant="compact" />
+        </div>
+      </details>
+    );
+  }
+
+  return (
+    <div className="space-y-1 px-1 py-0.5 text-xs text-slate-700">
+      <div className="flex min-h-6 items-center gap-2">
+        <Brain className="size-3.5 shrink-0 text-slate-500" />
+        <span className="shrink-0 font-medium text-slate-500">Thinking</span>
+        {timeLabel ? (
+          <span className="ml-auto shrink-0 text-[11px] text-slate-500">{timeLabel}</span>
+        ) : null}
+      </div>
+      <MarkdownRenderer markdown={content || "Thinking..."} variant="compact" />
+    </div>
+  );
+};
+
+type AssistantMessageProps = {
+  message: AgentChatMessage;
+  sessionSelectedModel: AgentModelSelection | null;
+  assistantAccentColor: string | undefined;
+};
+
+const AssistantMessage = ({
+  message,
+  sessionSelectedModel,
+  assistantAccentColor,
+}: AssistantMessageProps): ReactElement => {
+  const footer = getAssistantFooterData(message, sessionSelectedModel);
+  return (
+    <div className="space-y-2">
+      <MarkdownRenderer markdown={message.content} variant="document" />
+      {footer.infoParts.length > 0 ? (
+        <div className="flex items-center gap-2 text-xs text-slate-500">
+          <span
+            className="size-1.5 rounded-sm bg-amber-500"
+            style={assistantAccentColor ? { backgroundColor: assistantAccentColor } : undefined}
+          />
+          <span className="min-w-0 truncate">{footer.infoParts.join(" · ")}</span>
+        </div>
+      ) : null}
+    </div>
+  );
+};
+
+export type MessageBodyProps = {
+  message: AgentChatMessage;
+  sessionSelectedModel: AgentModelSelection | null;
+  assistantAccentColor: string | undefined;
+  timeLabel: string;
+  systemPromptBody: string;
+};
+
+export const MessageBody = ({
+  message,
+  sessionSelectedModel,
+  assistantAccentColor,
+  timeLabel,
+  systemPromptBody,
+}: MessageBodyProps): ReactElement => {
+  const meta = message.meta;
+
+  if (meta?.kind === "reasoning") {
+    return (
+      <ReasoningMessage
+        content={message.content}
+        completed={meta.completed}
+        timeLabel={timeLabel}
+      />
+    );
+  }
+
+  if (meta?.kind === "tool") {
+    if (isOdtWorkflowMutationToolName(meta.tool)) {
+      return <WorkflowToolMessage meta={meta} messageTimestamp={message.timestamp} />;
+    }
+    return (
+      <RegularToolMessage
+        meta={meta}
+        messageContent={message.content}
+        messageTimestamp={message.timestamp}
+        timeLabel={timeLabel}
+      />
+    );
+  }
+
+  if (meta?.kind === "subtask") {
+    return (
+      <div className="flex min-h-6 items-center gap-2 px-1 py-0.5 text-xs text-violet-700">
+        <MessageSquareQuote className="size-3.5 shrink-0 text-violet-500" />
+        <p className="shrink-0 font-medium">subagent {meta.agent}</p>
+        <p className="truncate text-violet-700/90">{meta.description}</p>
+        {timeLabel ? (
+          <span className="ml-auto shrink-0 text-[11px] text-slate-500">{timeLabel}</span>
+        ) : null}
+      </div>
+    );
+  }
+
+  if (message.role === "system" && message.content.startsWith(SYSTEM_PROMPT_PREFIX)) {
+    return (
+      <details className="rounded border border-slate-200 bg-slate-50/70">
+        <summary className="cursor-pointer px-2 py-1 text-xs font-medium text-slate-700">
+          Show system prompt
+        </summary>
+        <div className="border-t border-slate-200 px-2 py-2">
+          <MarkdownRenderer markdown={systemPromptBody} variant="compact" />
+        </div>
+      </details>
+    );
+  }
+
+  if (message.role === "user") {
+    return (
+      <>
+        <p className="whitespace-pre-wrap leading-6">{message.content}</p>
+        {timeLabel ? (
+          <p className="mt-2 text-right text-[11px] font-medium text-slate-500">{timeLabel}</p>
+        ) : null}
+      </>
+    );
+  }
+
+  if (message.role === "thinking" || message.role === "system") {
+    return <p className="whitespace-pre-wrap leading-6 text-slate-700">{message.content}</p>;
+  }
+
+  if (message.role === "assistant") {
+    return (
+      <AssistantMessage
+        message={message}
+        sessionSelectedModel={sessionSelectedModel}
+        assistantAccentColor={assistantAccentColor}
+      />
+    );
+  }
+
+  return <MarkdownRenderer markdown={message.content} variant="document" />;
+};
