+++ apps/desktop/src/components/features/agents/agent-chat/agent-chat-message-card-types.ts
@@ -1,0 +1,3 @@
+import type { AgentChatMessage } from "@/types/agent-orchestrator";
+
+export type ToolMeta = Extract<NonNullable<AgentChatMessage["meta"]>, { kind: "tool" }>;
